Move `Setting` type to `settings.d.ts`
